Big Update
Ironed out some issues with EJS and CSS not loading properly, and got the navbar to route properly.

Routes still need to be added, so that the pages actually go to where they're supposed to be. Some other issues persist, but a good amount of progress was made for the rest of the project